Fix: Instant feedback on multi tap tap-dance
For the simulation of the MT() functionality via tap-dance, the feedback
applies only once the dance is over. This relates to the `TAPPING_TERM`
duration, which gets reset after each following tap of the same key.
Therefore the dance gets interrupted manually after the second tap,
since no other state can be reached from here.

As example does the current key mapping use the simulated MT for 'J' key
in combination with the 'CTRL' modifier. The 'J' can get tapped
frequently in a row when using vim(like tools). While continue to tap
the key nothing happens. Just when the user stops to tap and the timer
with `TAPPING_TERM` erases, the final evaluation of the dance applies.
This bother the flow, since the user can not see that for example the
cursor moves line by line done (for the Vim example). Instead it was
jumping all lines at once after the dance was over.