Feature: Allow simple layer and modifier colors
Export the functionality to set the backlight color by the active layer and
modifier to the user space. This allows to configure the behavior with
two simple mappings within the key mapping.

Active modifiers take preference over active layers. If no modifier is
active or it has no color defined, the color of the layer gets used. If
the active layer has no color defined, no backlight is enabled.