Shift keys according to future layout
The 'B' and 'N' keys (on QWERTY) are hard to type and will move to the
thumb on the future custom build keyboard. Even they are not that nice
to type on the Ergodox this must be learned. This strengthened because
the GUI and TMUX keys must move to the outer thumbs (and have by this
commit). Removing the keycaps should help for a haptic feedback, since
already this commit message is hard.