Replace TMux prefix key with delete tap
The delete key got lost recently when reducing the total number of used
keys from 38 to 36. But it has turned out that it is quite annoying to
not have it.
The only key that was still kinda reserved and special was the TMux
prefix key. Not even the OS (GUI key) control key has it's own dedicated
key anymore. Therefore the TMux configuration got adopted to use prefix
that is easy to type with regular modifiers (i.e. `Ctrl-a` at the time
of writing). Therefore the key can now be used for the delete tap.
It makes sense to have the key on the right thumb cluster as it an
action that targets the right side of the cursor.