Add tapping term per key for dancing brackets
The most common case of dancing bracket keys is the single tap. With a
high tapping term the respond is very slow and the key input feels
laggy. So it is a balance to find a short enough time window to make
single taps feel fast enough, but still have time to allow triple taps
when needed.