Move layer toggle keys for i18n to inner column
The i18n layer is in fact just a collection of keycodes combined with
the altgr modifier (in fact right alt key). The difference is that it
puts the keys on better locations like `ä` on `a` etc.
The movement comes now because of two reasons. The long term goal is to
get rid of the fourth thumb key as the thumb is already taking over a
alpha key from the index finder (inner column, bottom row). That
combines with the fact that this is actually about a modifier key, it
makes sense to put into the middle row next to all the other modifiers.
As it is the least frequently used "modifier" it gets the only not home
row position in this row. Though it is still comfortable to type.