Remove right bracket keys for learning
The dancing brackets have been introduced to the left split. But the
right hand still offers the closing brackets. By muscle memory I
automatically use them as they just work (still). To practice the new
technique they must be removed. This leave atm a gap in the matrix. This
will be focus of future improvements.