allow to add user features conditionally
So far all user featues get statically added. This requires to "delete"
them when they are not used. This is a bit annoying. Though they are
recoverable via the repository history, it is not just as simple to
integrate it back to the new setup. This way it is a simple on/off
toggle.