reformat keymap layout to save line width
The pretty layout with both hands next to each other results into very
long lines. Editing this file without a wide screen and without ever
splitting the Vim windows is very annoying. Therfore it uses a layout
now which places both hands under each other.
It is planned to define a custom function for this to shrink it to the
34 used keys only.