split some layers into left and right hand
This has the advantage that it is easier to type keys from such a layer
in combination with a modifier from the other hand. At the moment this
requires to firt hold the modifier on the one hand until its gets
active, then hold the layer switch key on the same hand and finally
click the key on the other hand. With the changes it is bascially
possible to press all keys at once, or at least the first two and then
the third one. It feels way more comfortable and natural.
It achieves this by having one half completely set to being transparent
to the base layer and the other half has its keys set.
This change requires that it is never (or rarely) necessary to press
keys on the other half within the same layer. But this is a bad habit
anyway. It is not nice to press keys on the keyboard half where the
layer switch key is hold. In worst case it is necessary to switch the
hold on the other side. This is now bascially required as these are two
different layers. With the latest changes on the layer composition this
should be really rarely the case and is fine.