add new user feature for double space sentence
When the space key gets tapped two times in a row it will finish the
last sentence and start a new one. In details this means it will replace
the two spaces with a dot, a space and then activates shift as one-shot
modifier for the next character. This is very handy as while typing text
it just requires a very quick double tap two do a lot of work (including
a possible hold for shift).
This feature got heavily inspired by mobile phone keyboards. To enhance
this a little and allow to use the feature also while coding, the
application gets limited to the case that the last character before
the double space was a alphabetical character. While this restriction
might be annoying sometimes when a sentence ends with a closing
parentheses or similar, this is a compromise with way more often
occurring double spaces while coding.
This feature is under test. It might turn out to be more annoying than
helpful. For example it is cool that this feature works independently
everywhere, but it is also more stupid as it doesn't see the actual
text.