fix slash and question mark switch
I think I thought it would be handy to have only sentence related signs
in the base layer at the bottom right. But it is too annoying as it is
different from the default, but just a switch between layers/shift. As
it is really not important which way around it it and the effort between
both doesn't vary too much, just stick to the default.