recover dancing brackets instead of combos
Using the combos was an expriment. But it didn't just feel awkward
because it was new, it really felt uncomfortable and it was really hard
to hit them right. It was always interrupting the flow and very often it
did not work properly. The dancing brackets are way easier to use, even
in the worst case.